rename specs so they are picked up
